+++ packages/babel-plugin-fbt/src/__tests__/fbtAutoWrap-test.js
@@ -5,8 +5,14 @@
  * @format
  */
 
-const {payload, transform, withFbtRequireStatement} = require('../FbtTestUtil');
+const {
+  jsCodeSerializer,
+  snapshotTransform,
+  withFbtRequireStatement,
+} = require('../FbtTestUtil');
 const {TestUtil} = require('fb-babel-plugin-utils');
+
+expect.addSnapshotSerializer(jsCodeSerializer);
 
 const testData = {
   'should auto wrap a simple test with one level': {
@@ -15,32 +21,6 @@ const testData = {
         <link href="#">Your friends</link>
         liked your video
       </fbt>;`,
-    ),
-
-    output: withFbtRequireStatement(
-      `fbt._(
-        ${payload({
-          type: 'text',
-          jsfbt: '{=Your friends} liked your video',
-          desc: 'd',
-        })},
-        [
-          fbt._param(
-            '=Your friends',
-            React.createElement(
-              'link',
-              {href: '#'},
-              fbt._(
-                ${payload({
-                  type: 'text',
-                  jsfbt: 'Your friends',
-                  desc: 'In the phrase: "{=Your friends} liked your video"',
-                })},
-              ),
-            ),
-          ),
-        ],
-      );`,
     ),
   },
 
@@ -54,61 +34,18 @@ const testData = {
         your video
       </fbt>;`,
     ),
-
-    output: withFbtRequireStatement(
-      `fbt._(
-        ${payload({
-          type: 'text',
-          jsfbt: '{=Your friends liked} your video',
-          desc: 'd',
-        })},
-        [
-          fbt._param(
-            '=Your friends liked',
-            React.createElement(
-              Link,
-              {href: '#'},
-              fbt._(
-                ${payload({
-                  type: 'text',
-                  jsfbt: 'Your friends {=liked}',
-                  desc: 'In the phrase: "{=Your friends liked} your video"',
-                })},
-                [
-                  fbt._param(
-                    '=liked',
-                    React.createElement(
-                      'b',
-                      null,
-                      fbt._(
-                        ${payload({
-                          type: 'text',
-                          jsfbt: 'liked',
-                          desc:
-                            'In the phrase: "Your friends {=liked} your video"',
-                        })},
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      );`,
-    ),
   },
 };
 
 // TODO(T40113359) Re-enable once this test scenario is ready to be tested
 xdescribe('Test jsx auto-wrapping of implicit parameters', () =>
-  TestUtil.testSection(testData, transform));
+  TestUtil.testSection(testData, snapshotTransform, {matchSnapshot: true}));
 
 // TODO(T40113359) Re-enable once this test scenario is ready to be tested
 xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
   it('should wrap a single unwrapped <fbt> child and a string above', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=This is a nested">
@@ -130,7 +67,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <b>
@@ -146,7 +83,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap a single unwrapped <fbt> child and a string below', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div href="#">this is</div>
@@ -155,7 +92,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=this is">
@@ -174,7 +111,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap two unwrapped <fbt> children', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div>wrap once</div>
@@ -183,7 +120,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=wrap once">
@@ -208,7 +145,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap two unwrapped <fbt> children and 1 nested', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div>
@@ -221,7 +158,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=wrap once and also">
@@ -254,7 +191,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap an outer and inner child', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div href="#">
@@ -266,7 +203,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=this is a doubly">
@@ -292,7 +229,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap two children with one nested level', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div href="#">
@@ -305,7 +242,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=this is a doubly">
@@ -338,7 +275,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap a <fbt> child next to an explicit <fbt:param>', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="explicit param next to">
@@ -351,7 +288,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="explicit param next to">
@@ -374,7 +311,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap a <fbt> child nested in an explicit <fbt:param>', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="explicit fbt param">
@@ -389,7 +326,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="explicit fbt param">
@@ -417,7 +354,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
       'implicit <fbt:param> within it',
     () => {
       expect(
-        transform(
+        snapshotTransform(
           withFbtRequireStatement(
             `<fbt desc="d">
               outer string that should not appear in inner desc
@@ -433,7 +370,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
           ),
         ),
       ).toEqual(
-        transform(
+        snapshotTransform(
           withFbtRequireStatement(
             `<fbt desc="d">
               outer string that should not appear in inner desc
@@ -460,7 +397,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should work with multiple <fbt> calls in one file', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<div>
             <fbt desc="one">
@@ -475,7 +412,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<div>
             <fbt desc="one">
@@ -502,7 +439,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap two nested next to each other', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div href="#">
@@ -517,7 +454,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=one two">
@@ -556,7 +493,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap two nested next to each other with an extra level', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div href="#">
@@ -574,7 +511,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=one two test">
@@ -620,7 +557,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
 
   it('should wrap explicit params nested in implicit params with []', () => {
     expect(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <div>
@@ -655,7 +592,7 @@ xdescribe('Equality between auto-wrapped and manually wrapped params', () => {
         ),
       ),
     ).toEqual(
-      transform(
+      snapshotTransform(
         withFbtRequireStatement(
           `<fbt desc="d">
             <fbt:param name="=this is a test [to make sure that explicit params under an implicit node][and ones that are next to each other] under an implicit tag are wrapped with [ ]">